requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.1.3
django-ajax-datatable==4.5.0
django-crispy-forms==2.3
django-netfields==1.3.2
djangorestframework==3.15.2
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-api-keys==0.5.13
googleapis-common-protos==1.66.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
idna==3.10
netaddr==1.3.0
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
